pyproject.toml
[tool.poetry]
name = "rag_benchmarking"
version = "0.1.0"
description = ""
authors = [
    "Praveen Koka"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
mem0ai = ">=0.1.104,<0.2.0"
python-dotenv = ">=1.1.0,<2.0.0"
openai = ">=1.83.0,<2.0.0"
qdrant-client = ">=1.14.2,<2.0.0"
beir = ">=2.1.0,<3.0.0"
sentence-transformers = ">=4.1.0,<5.0.0"
requests = ">=2.32.3,<3.0.0"
azure-identity = ">=1.23.0,<2.0.0"
azure-search-documents = "==11.6.0b12"
tqdm = ">=4.67.1,<5.0.0"
datasets = "^3.6.0"
pinecone = "^7.0.2"
cohere = "^5.15.0"
google-genai = "^1.19.0"
scikit-learn = "^1.7.0"
numpy = "^2.2.6"
torch = "^2.7.1"
transformers = "^4.52.4"
openpyxl = "^3.1.5"
vertexai = "^1.71.1"
milvus = "^2.3.9"
pymilvus = "^2.5.10"
html2text = "^2025.4.15"
bs4 = "^0.0.2"
tiktoken = "^0.9.0"
inquirerpy = "^0.3.4"
accelerate = "^1.7.0"
jsonlines = "^4.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
